+++ prototipo/arquivos/personagens/player.py
@@ -1,51 +1,7 @@
 import pygame
 import math
 from equipamentos import *
-from abc import ABC, abstractmethod
-
-
-
-class Personagem(ABC, pygame.sprite.Sprite):
-    def __init__(self, vida, dano, arma: Arma, *groups):
-        super().__init__(*groups)
-        self.__vida = vida
-        self.__dano = dano
-        self.__arma = arma
-
-
-    @abstractmethod
-    def update(self):
-        pass
-
-    @property
-    def vida(self):
-        return self.__vida
-    @vida.setter
-    def vida(self,vida):
-        self.__vida = vida
-
-    @property
-    def dano(self):
-        return self.__dano
-    @dano.setter
-    def dano(self,dano):
-        self.__dano = dano
-
-    @property
-    def arma(self):
-        return self.__arma
-    @arma.setter
-    def arma(self, arma):
-        return self.__arma
-
-
-
-
-    def tomar_dano(self, qtdade_dano):
-        self.__vida = self.__vida - qtdade_dano
-
-
-
+from arquivos.personagens.personagem import Personagem
 
 class Player(Personagem):
     def __init__(self,vida ,dano ,arma, item: Item, *groups):
@@ -91,22 +47,3 @@ class Player(Personagem):
             return True
         else:
             return  False 
-
-
-
-class Enemy(Personagem):
-    def __init__(self, vida, dano, arma, imagem, *groups):
-        super().__init__(vida, dano, arma, *groups)
-
-        self.image = imagem
-
-        self.image = pygame.transform.scale(self.image, [20, 50])
-        self.rect = self.image.get_rect()
-        self.timer = 0
-
-
-    def update(self,*args):
-    # Comandos
-        self.timer += 0.003
-        self.rect.x = 512 + math.sin(self.timer) * 320
-
